solution test 2

+++ lab2-221220113/lab2/kernel/kernel/kvm.c
@@ -86,5 +86,5 @@ void loadUMain(void) {
 		*(unsigned char*)(elf + i) = *(unsigned char*)(elf + i + offset);
 	}
 
-	enterUserSpace(uMainEntry);
+	enterUserSpace((uint32_t)uMainEntry);
 }
